moved some funtions to xfunctions.c + some minor cleaning
