+++ src/globals.c
@@ -13,6 +13,12 @@
 // some structs
 //
 
+
+#ifndef _BE_POSIX
+#define CMD_LEN_MAX (PATH_MAX + ((NAME_MAX + 1) << 1))
+char len_buf[CMD_LEN_MAX] __attribute__ ((aligned));
+#endif
+
 struct usrvar_t *usr_var = (struct usrvar_t *)NULL;
 struct actions_t *usr_actions = (struct actions_t *)NULL;
 struct ws_t *ws = (struct ws_t *)NULL;
